Remove T commands and log related functions
