+++ librz/core/cmd_helps.h
@@ -37,8 +37,6 @@ extern const RzCmdDescHelp g_help;
 extern const RzCmdDescHelp i_help;
 extern const RzCmdDescHelp k_help;
 extern const RzCmdDescHelp ls_help;
-extern const RzCmdDescHelp j_help;
-extern const RzCmdDescHelp h_help;
 extern const RzCmdDescHelp L_help;
 extern const RzCmdDescHelp m_help;
 extern const RzCmdDescHelp o_help;
@@ -52,7 +50,6 @@ extern const RzCmdDescHelp colon_help;
 extern const RzCmdDescHelp rz_help;
 extern const RzCmdDescHelp s_help;
 extern const RzCmdDescHelp t_help;
-extern const RzCmdDescHelp T_help;
 extern const RzCmdDescHelp u_help;
 extern const RzCmdDescHelp pipein_help;
 extern const RzCmdDescHelp V_help;
