+++ librz/core/rtr.c
@@ -103,75 +103,6 @@ static void showcursor(RzCore *core, int x) {
 	rz_cons_flush ();
 }
 
-// TODO: rename /name to /nick or /so?
-// clone of textlog_chat () using rtrcmd()
-static void rtr_textlog_chat (RzCore *core, TextLog T) {
-	char prompt[64];
-	char buf[1024];
-	int lastmsg = 0;
-	const char *me = rz_config_get (core->config, "cfg.user");
-	char *ret, msg[1024];
-
-	eprintf ("Type '/help' for commands and ^D to quit:\n");
-	char *oldprompt = strdup (rz_line_singleton ()->prompt);
-	snprintf (prompt, sizeof (prompt) - 1, "[%s]> ", me);
-	rz_line_set_prompt (prompt);
-	ret = rtrcmd (T, msg);
-	for (;;) {
-		if (lastmsg >= 0) {
-			snprintf (msg, sizeof (msg) - 1, "T %d", lastmsg);
-		} else {
-			strcpy (msg, "T");
-		}
-		ret = rtrcmd (T, msg);
-		rz_cons_println (ret);
-		free (ret);
-		ret = rtrcmd (T, "Tl");
-		lastmsg = atoi (ret)-1;
-		free (ret);
-		if (rz_cons_fgets (buf, sizeof (buf), 0, NULL) < 0) {
-			goto beach;
-		}
-		if (!*buf) {
-			continue;
-		}
-		if (!strcmp (buf, "/help")) {
-			eprintf ("/quit           quit the chat (same as ^D)\n");
-			eprintf ("/nick <nick>    set cfg.user nick name\n");
-			eprintf ("/log            show full log\n");
-			eprintf ("/clear          clear text log messages\n");
-		} else if (!strncmp (buf, "/nick ", 6)) {
-			snprintf (msg, sizeof (msg) - 1, "* '%s' is now known as '%s'", me, buf+6);
-			rz_cons_println (msg);
-			rz_core_log_add (core, msg);
-			rz_config_set (core->config, "cfg.user", buf+6);
-			me = rz_config_get (core->config, "cfg.user");
-			snprintf (prompt, sizeof (prompt) - 1, "[%s]> ", me);
-			rz_line_set_prompt (prompt);
-		} else if (!strcmp (buf, "/log")) {
-			char *ret = rtrcmd (T, "T");
-			if (ret) {
-				rz_cons_println (ret);
-				free (ret);
-			}
-		} else if (!strcmp (buf, "/clear")) {
-			//rz_core_log_del (core, 0);
-			free (rtrcmd (T, "T-"));
-		} else if (!strcmp (buf, "/quit")) {
-			goto beach;
-		} else if (*buf=='/') {
-			eprintf ("Unknown command: %s\n", buf);
-		} else {
-			char *cmd = rz_str_newf ("T [%s] %s", me, buf);
-			free (rtrcmd (T, cmd));
-			free (cmd);
-		}
-	}
-beach:
-	rz_line_set_prompt (oldprompt);
-	free (oldprompt);
-}
-
 RZ_API int rz_core_rtr_http_stop(RzCore *u) {
 	RzCore *core = (RzCore*)u;
 	const int timeout = 1; // 1 second
